feat: adapted to the latest patch structure
