+++ vllm_ascend/ops/rotary_embedding.py
@@ -23,6 +23,7 @@ from vllm.model_executor.layers.rotary_embedding import (
     DeepseekScalingRotaryEmbedding, RotaryEmbedding)
 
 from vllm_ascend.utils import enable_custom_op
+from vllm_ascend.utils import is_310p
 
 
 def custom_rotary_embedding_enabled(query, neox_style, head_size):
@@ -48,7 +49,8 @@ def rope_forward_oot(
     if is_neox_style_override is not None:
         neox_style = is_neox_style_override
     # adopt custom kernel path for rotary_embedding
-    if custom_rotary_embedding_enabled(query, neox_style, self.head_size):
+    if custom_rotary_embedding_enabled(query, neox_style,
+                                       self.head_size) and not is_310p():
         query, key = torch.ops._C.rotary_embedding(
             positions,
             query,
