+++ vllm_ascend/platform.py
@@ -28,9 +28,7 @@ from vllm.logger import logger
 from vllm.platforms import Platform, PlatformEnum
 
 from vllm_ascend.ascend_config import check_ascend_config, init_ascend_config
-from vllm_ascend.utils import (ASCEND_QUATIZATION_METHOD,
-                               communication_adaptation_310p, is_310p,
-                               update_aclgraph_sizes)
+from vllm_ascend.utils import ASCEND_QUATIZATION_METHOD, update_aclgraph_sizes
 
 if TYPE_CHECKING:
     from vllm.config import ModelConfig, VllmConfig
@@ -76,10 +74,6 @@ class NPUPlatform(Platform):
 
         from vllm_ascend.quantization.quant_config import \
             AscendQuantConfig  # noqa: F401
-
-        if is_310p():
-            logger.info("patch torch communication while using Ascend 310P")
-            communication_adaptation_310p()
 
     @classmethod
     def get_device_capability(cls, device_id: int = 0):
