+++ vllm_ascend/utils.py
@@ -141,69 +141,6 @@ def aligned_16(tensor: torch.Tensor):
     return new_tensor
 
 
-def communication_adaptation_310p():
-
-    def broadcast310p(tensor, src, group=None, async_op=False):
-        rank = torch.distributed.get_rank(group)
-        world_size = torch.distributed.get_world_size(group)
-        tensor_list = [torch.empty_like(tensor) for _ in range(world_size)]
-        tensor_list[rank] = tensor
-        torch.distributed.all_gather(tensor_list, tensor, group=group)
-        tensor[...] = tensor_list[src]
-        if async_op:
-            return NullHandle()
-        else:
-            return None
-
-    torch.distributed.broadcast = broadcast310p
-    torch.distributed.distributed_c10d.broadcast = broadcast310p
-
-    def all_reduce_wrapper_310p(fn):
-
-        def all_reduce(
-            tensor,
-            op=torch.distributed.ReduceOp.SUM,
-            group=None,
-            async_op=False,
-        ):
-            if tensor.dtype != torch.int64:
-                return fn(tensor, op, group, async_op)
-            rank = torch.distributed.get_rank(group)
-            world_size = torch.distributed.get_world_size(group)
-            tensor_list = [torch.empty_like(tensor) for _ in range(world_size)]
-            tensor_list[rank] = tensor
-            torch.distributed.all_gather(tensor_list, tensor, group=group)
-            if op == torch.distributed.ReduceOp.SUM:
-                return torch.stack(tensor_list).sum(0)
-            elif op == torch.distributed.ReduceOp.MAX:
-                return torch.tensor(
-                    torch.stack(tensor_list).cpu().numpy().max(0),
-                    device=tensor.device,
-                )
-            else:
-                raise RuntimeError(f"not implement op {op}")
-
-        return all_reduce
-
-    torch.distributed.all_reduce = all_reduce_wrapper_310p(
-        torch.distributed.all_reduce)
-    torch.distributed.distributed_c10d.all_reduce = all_reduce_wrapper_310p(
-        torch.distributed.distributed_c10d.all_reduce)
-
-    def reduce_scatter_310p(output_tensor, input_tensor, group=None):
-        rank = torch.distributed.get_rank(group)
-        world_size = torch.distributed.get_world_size(group)
-        torch.distributed.all_reduce(input_tensor,
-                                     torch.distributed.ReduceOp.SUM,
-                                     group,
-                                     async_op=False)
-        interval = input_tensor.shape[0] // world_size
-        output_tensor[:] = input_tensor[rank * interval:(rank + 1) * interval]
-
-    torch.distributed._reduce_scatter_base = reduce_scatter_310p
-    torch.distributed.distributed_c10d._reduce_scatter_base = reduce_scatter_310p
-
-
 def try_register_lib(lib_name: str, lib_info: str = ""):
     import importlib
     import importlib.util
